# docs/swagger.yaml
consumes:
- application/json
definitions:
...
  github_com_etkecc_mrs_internal_model.IndexStats:
    properties:
      discovery:
        $ref: '#/definitions/github_com_etkecc_mrs_internal_model.IndexStatsTime'
      indexing:
        $ref: '#/definitions/github_com_etkecc_mrs_internal_model.IndexStatsTime'
      parsing:
        $ref: '#/definitions/github_com_etkecc_mrs_internal_model.IndexStatsTime'
      rooms:
        $ref: '#/definitions/github_com_etkecc_mrs_internal_model.IndexStatsRooms'
      servers:
        $ref: '#/definitions/github_com_etkecc_mrs_internal_model.IndexStatsServers'
    type: object
  github_com_etkecc_mrs_internal_model.IndexStatsRooms:
    properties:
      banned:
        type: integer
      indexed:
        type: integer
      parsed:
        type: integer
      reported:
        type: integer
    type: object
  github_com_etkecc_mrs_internal_model.IndexStatsServers:
    properties:
      blocked:
        type: integer
      indexable:
        type: integer
      online:
        type: integer
      software:
        additionalProperties:
          type: integer
        type: object
    type: object
  github_com_etkecc_mrs_internal_model.IndexStatsTime:
    properties:
      finished_at:
        type: string
      started_at:
        type: string
    type: object
  github_com_etkecc_mrs_internal_model.MatrixError:
    properties:
      errcode:
        description: Matrix error code, e.g M_UNAUTHORIZED
        type: string
      error:
        description: Matrix error message
        type: string
    type: object
  github_com_etkecc_mrs_internal_model.MatrixRoom:
    properties:
      avatar_url:
        type: string
      avatar_url_http:
        type: string
      canonical_alias:
        type: string
      email:
        type: string
      guest_can_join:
        type: boolean
      join_rule:
        type: string
      language:
        type: string
      name:
        type: string
      num_joined_members:
        type: integer
      parsed_at:
        type: string
      room_id:
        type: string
      room_type:
        type: string
      server:
        description: Parsed (custom) fields
        type: string
      servers:
        items:
          type: string
        type: array
      topic:
        type: string
      world_readable:
        type: boolean
    type: object
  github_com_etkecc_mrs_internal_model.MatrixServer:
    properties:
      checked_at:
        type: string
      contacts:
        allOf:
        - $ref: '#/definitions/github_com_etkecc_mrs_internal_model.MatrixServerContacts'
        description: Contacts as per MSC1929
      indexable:
        description: Is the server published the public room directory over federation?
        type: boolean
      name:
        description: ServerName, as per spec, e.g., "example.com"
        type: string
      online:
        description: Is the server online and federating?
        type: boolean
      online_at:
        description: |-
          OnlineAt is the prune clock: last seen online, the sole basis for the 30d offline delete. Never bumped on an offline dial.
          CheckedAt is the backoff clock: last dial attempt, bumped every dial. Merge them and dead servers reset the prune clock and go immortal.
        type: string
      software:
        description: Software running on the server, e.g., Synapse, Dendrite
        type: string
      url:
        description: Server-Server API URL, e.g., "https://example.com:8448"
        type: string
      version:
        description: Version of the software, e.g., 1.0.0
        type: string
    type: object
  github_com_etkecc_mrs_internal_model.MatrixServerContacts:
    properties:
      emails:
        items:
          type: string
        type: array
      mxids:
        items:
          type: string
        type: array
      url:
        type: string
    type: object
...
info:
  contact: {}
  description: Matrix Rooms Search crawls the Matrix federation and serves public-room
    search. It also answers part of the Matrix federation and client-server APIs for
    the rooms it has indexed. Where our implementation diverges from the spec, that
    divergence is documented here on purpose, not hidden. matrixrooms.info is a public
    demo running MRS.
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0
  title: Matrix Rooms Search API
  version: 0.0.0
paths:
  /-/discover:
    post:
      description: 'Kicks off a discovery pass and returns 201 immediately. Fire-and-forget:
        the crawl runs in the background, this does not wait for it.'
      produces:
      - application/json
      responses:
        "201":
          description: Discovery started in the background
      security:
      - AdminAuth: []
      summary: Trigger discovery
      tags:
      - admin
  /-/full:
    post:
      description: Runs a full cycle (discovery, then parsing) in one go and returns
        201 immediately. Fire-and-forget, runs in the background. This is what a periodic
        cron/timer should hit to keep the index fresh.
      produces:
      - application/json
      responses:
        "201":
          description: Full cycle started in the background
      security:
      - AdminAuth: []
      summary: Trigger a full cycle
      tags:
      - admin
  /-/parse:
    post:
      description: Kicks off a room-parsing pass and returns 201 immediately. Fire-and-forget,
        runs in the background.
      produces:
      - application/json
      responses:
        "201":
          description: Parsing started in the background
      security:
      - AdminAuth: []
      summary: Trigger parsing
      tags:
      - admin
  /-/reindex:
    post:
      description: Rebuilds the search index from what is already crawled and returns
        201 immediately. Fire-and-forget, runs in the background.
      produces:
      - application/json
      responses:
        "201":
          description: Reindex started in the background
      security:
      - AdminAuth: []
      summary: Trigger reindex
      tags:
      - admin
  /-/status:
    get:
      description: 'Full crawler and index statistics: server and room counts, plus
        the timing of the last discovery, parsing, and indexing passes. The admin-side
        twin of the public /stats, with more detail.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.IndexStats'
      security:
      - AdminAuth: []
      summary: Index status
      tags:
      - admin
...
  /_matrix/media/r0/thumbnail/{name}/{id}:
    get:
      description: Serves a media thumbnail. Tries the unauthenticated client-server
        media API first (faster), then falls back to the authenticated federation
        media API. Streams the image on a hit, or 204 when there is nothing to serve.
        Standard Matrix thumbnail query params (width, height, method) are passed
        straight through.
      parameters:
      - description: Media origin server name
        in: path
        name: name
        required: true
        type: string
      - description: Media ID
        in: path
        name: id
        required: true
        type: string
      - description: Thumbnail width
        in: query
        name: width
        type: integer
      - description: Thumbnail height
        in: query
        name: height
        type: integer
      - description: 'Resize method: crop or scale'
        in: query
        name: method
        type: string
      produces:
      - image/jpeg
      - image/png
      responses:
        "200":
          description: Thumbnail image
          schema:
            type: file
        "204":
          description: No thumbnail available
      summary: Media thumbnail
      tags:
      - media
  /_matrix/media/v3/thumbnail/{name}/{id}:
    get:
      description: Serves a media thumbnail. Tries the unauthenticated client-server
        media API first (faster), then falls back to the authenticated federation
        media API. Streams the image on a hit, or 204 when there is nothing to serve.
        Standard Matrix thumbnail query params (width, height, method) are passed
        straight through.
      parameters:
      - description: Media origin server name
        in: path
        name: name
        required: true
        type: string
      - description: Media ID
        in: path
        name: id
        required: true
        type: string
      - description: Thumbnail width
        in: query
        name: width
        type: integer
      - description: Thumbnail height
        in: query
        name: height
        type: integer
      - description: 'Resize method: crop or scale'
        in: query
        name: method
        type: string
      produces:
      - image/jpeg
      - image/png
      responses:
        "200":
          description: Thumbnail image
          schema:
            type: file
        "204":
          description: No thumbnail available
      summary: Media thumbnail
      tags:
      - media
  /avatar/{name}/{id}:
    get:
      description: Serves a media thumbnail. Tries the unauthenticated client-server
        media API first (faster), then falls back to the authenticated federation
        media API. Streams the image on a hit, or 204 when there is nothing to serve.
        Standard Matrix thumbnail query params (width, height, method) are passed
        straight through.
      parameters:
      - description: Media origin server name
        in: path
        name: name
        required: true
        type: string
      - description: Media ID
        in: path
        name: id
        required: true
        type: string
      - description: Thumbnail width
        in: query
        name: width
        type: integer
      - description: Thumbnail height
        in: query
        name: height
        type: integer
      - description: 'Resize method: crop or scale'
        in: query
        name: method
        type: string
      produces:
      - image/jpeg
      - image/png
      responses:
        "200":
          description: Thumbnail image
          schema:
            type: file
        "204":
          description: No thumbnail available
      summary: Media thumbnail
      tags:
      - media
  /catalog/rooms:
    get:
      description: Every indexed room as a room-ID to alias map. Big, authenticated,
        and exactly as heavy as it sounds.
      produces:
      - application/json
      responses:
        "200":
          description: Room ID to alias
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CatalogAuth: []
      summary: All rooms
      tags:
      - catalog
  /catalog/servers:
    get:
      description: The servers the crawler currently considers online.
      produces:
      - application/json
      responses:
        "200":
          description: Online server names
          schema:
            items:
              type: string
            type: array
      security:
      - CatalogAuth: []
      summary: Online servers
      tags:
      - catalog
  /catalog/servers/objects:
    get:
      description: Online servers with their full crawl records, keyed by server name.
      produces:
      - application/json
      responses:
        "200":
          description: Server name to crawl record
          schema:
            additionalProperties:
              $ref: '#/definitions/github_com_etkecc_mrs_internal_model.MatrixServer'
            type: object
      security:
      - CatalogAuth: []
      summary: Online servers with details
      tags:
      - catalog
  /discover/{name}:
    post:
      description: 'Queues a single server for discovery. Auth is optional: send Discovery
        credentials to skip the rate limit, or call anonymously and take the limit.
        We dial the server before answering, so the status code tells you what actually
        happened: 201 if it is reachable and freshly added, 208 if we already knew
        it, 422 if it would not answer (we record it offline rather than re-dial it
        on every retry).'
      parameters:
      - description: Server name to add
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Server was reachable and queued
        "208":
          description: Server already known, nothing to do
        "401":
          description: Invalid Discovery credentials (only if you send them)
        "422":
          description: Server could not be reached, nothing added
        "429":
          description: Rate limited (anonymous callers)
      summary: Add a server
      tags:
      - discovery
  /discover/bulk:
    post:
      consumes:
      - application/json
      description: 'Queues a batch of servers for discovery. Requires Discovery credentials.
        Fire-and-forget: we take the JSON array, return 202 immediately, and discover
        them in the background.'
      parameters:
      - description: Server names to add
        in: body
        name: request
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "202":
          description: Batch accepted for background discovery
        "400":
          description: Malformed request body (not a JSON array of server names)
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.MatrixError'
        "401":
          description: Invalid Discovery credentials
      security:
      - DiscoveryAuth: []
      summary: Add servers in bulk
      tags:
      - discovery
...
  /room/{room_id_or_alias}:
    get:
      description: 'Room preview by ID or alias: like the Matrix client-server room
        preview, but enriched with everything MRS knows about a room (language and
        so on). We try our own index first, then fall back to a live MSC3266 summary,
        and when that fallback fires we set the `X-MRS-MSC3266: true` response header
        so you can tell. Marked EXPERIMENT in the source, so treat the shape as not
        yet frozen.'
      parameters:
      - description: Room ID or alias
        in: path
        name: room_id_or_alias
        required: true
        type: string
      - description: Server to try for the MSC3266 fallback if we do not have the
          room indexed
        in: query
        name: via
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Room data
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.MatrixRoom'
        "404":
          description: Room not found anywhere we looked
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.MatrixError'
        "500":
          description: Internal error reading the room
          schema:
            $ref: '#/definitions/github_com_etkecc_mrs_internal_model.MatrixError'
      summary: Room preview
      tags:
      - catalog
...
produces:
- application/json
schemes:
- https
securityDefinitions:
  AdminAuth:
    type: basic
  CatalogAuth:
    type: basic
  DiscoveryAuth:
    type: basic
  FederationAuth:
    description: 'Matrix federation signature auth: an `Authorization: X-Matrix origin=...,key=...,sig=...`
      header carrying an ed25519 signature over the request, which we verify against
      the origin server''s published keys.'
    in: header
    name: Authorization
    type: apiKey
  MetricsAuth:
    type: basic
  ModerationAuth:
    type: basic
swagger: "2.0"